[refactor] reject revision to revision controller
