+++ Backend/Controllers/RoastersController.cs
@@ -66,13 +66,15 @@ public class RoastersController(IRoasterService roastersService) : BaseApiContro
   public async Task<IActionResult> GetRevisionDiff([FromQuery] int revisionId1, [FromQuery] int revisionId2, int roasterId)
     => (await roastersService.GetRoasterRevisionDiffAsync(revisionId1, revisionId2, roasterId, User)).ToActionResult();
 
-  [Authorize(Policy = "RequireModeratorRole")]
-  [HttpPost("{roasterId:int}/revisions/{revisionId:int}/approve")]
-  public async Task<IActionResult> ApproveRoasterRevision(int roasterId, int revisionId)
-    => (await roastersService.ApproveRoasterRevisionAsync(roasterId, revisionId)).ToActionResult();
+  // [Authorize(Policy = "RequireModeratorRole")]
+  // [HttpPost("{roasterId:int}/revisions/{revisionId:int}/approve")]
+  // // if theres no parent, then roasterId will be null since it hasnt been created yet
+  // // just make a 1 input overload for ApproveRoasterRevisionAsync
+  // public async Task<IActionResult> ApproveRoasterRevision(int roasterId, int revisionId)
+  //   => (await roastersService.ApproveRoasterRevisionAsync(roasterId, revisionId)).ToActionResult();
 
-  [Authorize(Policy = "RequireModeratorRole")]
-  [HttpPost("{roasterId:int}/revisions/{revisionId:int}/reject")]
-  public async Task<IActionResult> RejectRoasterRevision(int roasterId, int revisionId)
-    => (await roastersService.RejectRoasterRevisionAsync(roasterId, revisionId)).ToActionResult();
+  // [Authorize(Policy = "RequireModeratorRole")]
+  // [HttpPost("{roasterId:int}/revisions/{revisionId:int}/approve")]
+  // public async Task<IActionResult> ApproveRoasterRevision(int roasterId, int revisionId)
+  //   => (await roastersService.ApproveRoasterRevisionAsync(roasterId, revisionId)).ToActionResult();
 }
